refactor: Move ChessBoard component to component package.
