+++ app/src/main/java/no/usn/mob3000/ui/screens/chess/PlayScreen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import no.usn.mob3000.R
 import no.usn.mob3000.ui.components.base.Viewport
-import no.usn.mob3000.ui.screens.chess.game.ChessBoard
+import no.usn.mob3000.ui.components.game.board.ChessBoard
 import no.usn.mob3000.ui.theme.DefaultButton
 import no.usn.mob3000.ui.theme.DefaultListItemBackground
